+++ sdk/typescript/src/index.ts
@@ -1,1 +1,5 @@
 export * from './models/index.js';
+export {Hypervisor} from './hypervisor.js';
+export {VM} from './vm.js';
+export {HttpError, type CallOptions, type TransportOptions} from './transport.js';
+export type * from './options.js';